+++ streams/src/test/java/org/apache/kafka/streams/StreamsConfigTest.java
@@ -44,6 +44,7 @@ import org.apache.kafka.streams.processor.internals.RecordCollectorTest;
 import org.apache.kafka.streams.processor.internals.StreamsPartitionAssignor;
 import org.apache.kafka.streams.state.BuiltInDslStoreSuppliers;
 
+import org.apache.kafka.streams.utils.TestUtils.CountingProcessorWrapper;
 import org.apache.log4j.Level;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -1230,13 +1231,13 @@ public class StreamsConfigTest {
 
     @Test
     public void shouldAllowConfiguringProcessorWrapperWithClass() {
-        props.put(StreamsConfig.PROCESSOR_WRAPPER_CLASS_CONFIG, NoOpProcessorWrapper.class);
+        props.put(StreamsConfig.PROCESSOR_WRAPPER_CLASS_CONFIG, CountingProcessorWrapper.class);
         new StreamsConfig(props);
     }
 
     @Test
     public void shouldAllowConfiguringProcessorWrapperWithClassName() {
-        props.put(StreamsConfig.PROCESSOR_WRAPPER_CLASS_CONFIG, NoOpProcessorWrapper.class.getName());
+        props.put(StreamsConfig.PROCESSOR_WRAPPER_CLASS_CONFIG, CountingProcessorWrapper.class.getName());
         new StreamsConfig(props);
     }
